rename to resolveProps
