+++ packages/mui-material/src/Button/Button.js
@@ -1,7 +1,7 @@
 import * as React from 'react';
 import PropTypes from 'prop-types';
 import clsx from 'clsx';
-import { internal_mergeProps as mergeProps } from '@mui/utils';
+import { internal_resolveProps as resolveProps } from '@mui/utils';
 import { unstable_composeClasses as composeClasses } from '@mui/base';
 import { alpha } from '@mui/system';
 import styled, { rootShouldForwardProp } from '../styles/styled';
@@ -282,8 +282,8 @@ const ButtonEndIcon = styled('span', {
 const Button = React.forwardRef(function Button(inProps, ref) {
   // props priority: `inProps` > `contextProps` > `themeDefaultProps`
   const contextProps = React.useContext(ButtonGroupContext);
-  const mergedProps = mergeProps(contextProps, inProps);
-  const props = useThemeProps({ props: mergedProps, name: 'MuiButton' });
+  const resolvedProps = resolveProps(contextProps, inProps);
+  const props = useThemeProps({ props: resolvedProps, name: 'MuiButton' });
   const {
     children,
     color = 'primary',
