+++ packages/mui-utils/src/index.ts
@@ -33,4 +33,4 @@ export {
 export { default as usePreviousProps } from './usePreviousProps';
 export { default as visuallyHidden } from './visuallyHidden';
 export { default as integerPropType } from './integerPropType';
-export { default as internal_mergeProps } from './mergeProps';
+export { default as internal_resolveProps } from './resolveProps';
