+++ packages/mui-utils/src/resolveProps.test.ts
@@ -1,46 +1,46 @@
 import { expect } from 'chai';
-import mergeProps from './mergeProps';
+import resolveProps from './resolveProps';
 
-describe('mergeProps', () => {
+describe('resolveProps', () => {
   it('use default props if no props', () => {
-    expect(mergeProps({ foo: 'foo' }, {})).to.deep.equal({
+    expect(resolveProps({ foo: 'foo' }, {})).to.deep.equal({
       foo: 'foo',
     });
   });
 
   it('use props if defined', () => {
-    expect(mergeProps({ foo: 'foo' }, { foo: 'bar' })).to.deep.equal({
+    expect(resolveProps({ foo: 'foo' }, { foo: 'bar' })).to.deep.equal({
       foo: 'bar',
     });
   });
 
   it('merge extra props', () => {
-    expect(mergeProps({ foo: 'foo' }, { foo: 'bar', bar: 'bar' })).to.deep.equal({
+    expect(resolveProps({ foo: 'foo' }, { foo: 'bar', bar: 'bar' })).to.deep.equal({
       foo: 'bar',
       bar: 'bar',
     });
   });
 
   it('use default props if prop value is undefined', () => {
-    expect(mergeProps({ foo: 'foo' }, { foo: undefined })).to.deep.equal({
+    expect(resolveProps({ foo: 'foo' }, { foo: undefined })).to.deep.equal({
       foo: 'foo',
     });
   });
 
   it('use props if default value is undefined', () => {
-    expect(mergeProps({ foo: undefined }, { foo: 'bar' })).to.deep.equal({
+    expect(resolveProps({ foo: undefined }, { foo: 'bar' })).to.deep.equal({
       foo: 'bar',
     });
   });
 
   it('null is a considered a valid value', () => {
-    expect(mergeProps({ foo: 'foo' }, { foo: null })).to.deep.equal({
+    expect(resolveProps({ foo: 'foo' }, { foo: null })).to.deep.equal({
       foo: null,
     });
   });
 
   it('"" is a considered a valid value', () => {
-    expect(mergeProps({ foo: 'foo' }, { foo: '' })).to.deep.equal({
+    expect(resolveProps({ foo: 'foo' }, { foo: '' })).to.deep.equal({
       foo: '',
     });
   });
